Fixed "ERROR:  integer out of range"
There is no reason in limiting the variable to integers so we've upgraded to bigint.

+++ pg_sequence_to_identity/pg_sti.py
@@ -110,7 +110,7 @@ def _migrate_to_identity(schema, table, column, sql_only=False):
     SQL="""do
 $F$
     declare
-        new_max int;
+        new_max bigint;
         new_sequence text;
     begin
         select last_value+1 into new_max FROM %s ;
